requirements.txt
asgiref==3.11.1
Django==6.0.3
pillow==12.1.1
python-dotenv==1.2.2
sqlparse==0.5.5
